+++ packages/gatsby/src/query/query-compiler.js
@@ -31,7 +31,7 @@ const {
   VariablesInAllowedPositionRule,
 } = require(`graphql`)
 
-const getGatsbyDependents = require(`../utils/gatsby-dependents`)
+import { getGatsbyDependents } from "../utils/gatsby-dependents"
 const { store } = require(`../redux`)
 import * as actions from "../redux/actions/internal"
 const { default: FileParser } = require(`./file-parser`)
@@ -46,6 +46,7 @@ const {
   default: errorParser,
   locInGraphQlToLocInFile,
 } = require(`./error-parser`)
+const { websocketManager } = require(`../utils/websocket-manager`)
 
 const overlayErrorID = `graphql-compiler`