+++ packages/gatsby/src/redux/run-sift.js
@@ -1,9 +1,9 @@
 // @flow
 const { default: sift } = require(`sift`)
-const _ = require(`lodash`)
 const { prepareRegex } = require(`../utils/prepare-regex`)
 const { makeRe } = require(`micromatch`)
 import { getValueAt } from "../utils/get-value-at"
+import _ from "lodash"
 const {
   toDottedFields,
   objectToDottedField,
@@ -106,7 +106,7 @@ function handleMany(siftArgs, nodes) {
  * @param {Array<string>} chain Note: `$eq` is assumed to be the leaf prop here
  * @param {boolean | number | string} targetValue chain.a.b.$eq === targetValue
  * @param {Array<string>} nodeTypeNames
- * @param {undefined | Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
+ * @param {Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
  * @returns {Array<Node> | undefined}
  */
 const runFlatFilterWithoutSift = (
@@ -129,11 +129,6 @@ const runFlatFilterWithoutSift = (
   // There are also cases (and tests) where id exists with a different type
   if (!nodesByKeyValue) {
     return undefined
-  }
-
-  if (chain.join(`,`) === `id`) {
-    // The `id` key is not indexed in Sets (because why) so don't spread it
-    return [nodesByKeyValue]
   }
 
   // In all other cases this must be a non-empty Set because the indexing
@@ -165,17 +160,19 @@ const runFilterAndSort = (args: Object) => {
     firstOnly = false,
     nodeTypeNames,
     typedKeyValueIndexes,
+    stats,
   } = args
 
-  let result = applyFilters(
+  const result = applyFilters(
     filter,
     firstOnly,
     nodeTypeNames,
     typedKeyValueIndexes,
-    resolvedFields
+    resolvedFields,
+    stats
   )
 
-  return sortNodes(result, sort, resolvedFields)
+  return sortNodes(result, sort, resolvedFields, stats)
 }
 
 exports.runSift = runFilterAndSort
@@ -198,7 +195,8 @@ const applyFilters = (
   firstOnly,
   nodeTypeNames,
   typedKeyValueIndexes,
-  resolvedFields
+  resolvedFields,
+  stats
 ) => {
   const filters = filterFields
     ? prefixResolvedFields(
@@ -207,8 +205,26 @@ const applyFilters = (
       )
     : []
 
+  if (stats) {
+    filters.forEach(filter => {
+      const filterStats = filterToStats(filter)
+      const comparatorPath = filterStats.comparatorPath.join(`.`)
+      stats.comparatorsUsed.set(
+        comparatorPath,
+        (stats.comparatorsUsed.get(comparatorPath) || 0) + 1
+      )
+      stats.uniqueFilterPaths.add(filterStats.filterPath.join(`.`))
+    })
+    if (filters.length > 1) {
+      stats.totalNonSingleFilters++
+    }
+  }
+
   const result = filterWithoutSift(filters, nodeTypeNames, typedKeyValueIndexes)
   if (result) {
+    if (stats) {
+      stats.totalIndexHits++
+    }
     if (firstOnly) {
       return result.slice(0, 1)
     }
@@ -218,6 +234,21 @@ const applyFilters = (
   return filterWithSift(filters, firstOnly, nodeTypeNames, resolvedFields)
 }
 
+const filterToStats = (filter, filterPath = [], comparatorPath = []) => {
+  if (filter.type === `elemMatch`) {
+    return filterToStats(
+      filter.nestedQuery,
+      filterPath.concat(filter.path),
+      comparatorPath.concat([`elemMatch`])
+    )
+  } else {
+    return {
+      filterPath: filterPath.concat(filter.path),
+      comparatorPath: comparatorPath.concat(filter.query.comparator),
+    }
+  }
+}
+
 /**
  * Check if the filter is "flat" (single leaf) and an "$eq". If so, uses custom
  * indexes based on filter and types and returns any result it finds.
@@ -225,7 +256,7 @@ const applyFilters = (
  *
  * @param {Object} filter Resolved. (Should be checked by caller to exist)
  * @param {Array<string>} nodeTypeNames
- * @param {undefined | Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
+ * @param {Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
  * @returns {Array|undefined} Collection of results
  */
 const filterWithoutSift = (filters, nodeTypeNames, typedKeyValueIndexes) => {
@@ -263,9 +294,9 @@ exports.filterWithoutSift = filterWithoutSift
  *   will be limited to 1 if `firstOnly` is true
  */
 const filterWithSift = (filter, firstOnly, nodeTypeNames, resolvedFields) => {
-  let nodes = []
-
-  nodeTypeNames.forEach(typeName => addResolvedNodes(typeName, nodes))
+  const nodes = [].concat(
+    ...nodeTypeNames.map(typeName => addResolvedNodes(typeName))
+  )
 
   return _runSiftOnNodes(
     nodes,
@@ -364,7 +395,7 @@ const _runSiftOnNodes = (
  * @param resolvedFields
  * @returns {Array<Node> | undefined | null} Same as input, except sorted
  */
-const sortNodes = (nodes, sort, resolvedFields) => {
+const sortNodes = (nodes, sort, resolvedFields, stats) => {
   if (!sort || nodes?.length <= 1) {
     return nodes
   }
@@ -372,19 +403,24 @@ const sortNodes = (nodes, sort, resolvedFields) => {
   // create functions that return the item to compare on
   const dottedFields = objectToDottedField(resolvedFields)
   const dottedFieldKeys = Object.keys(dottedFields)
-  const sortFields = sort.fields
-    .map(field => {
-      if (
-        dottedFields[field] ||
-        dottedFieldKeys.some(key => field.startsWith(key))
-      ) {
-        return `__gatsby_resolved.${field}`
-      } else {
-        return field
-      }
-    })
-    .map(field => v => getValueAt(v, field))
+  const sortFields = sort.fields.map(field => {
+    if (
+      dottedFields[field] ||
+      dottedFieldKeys.some(key => field.startsWith(key))
+    ) {
+      return `__gatsby_resolved.${field}`
+    } else {
+      return field
+    }
+  })
+  const sortFns = sortFields.map(field => v => getValueAt(v, field))
   const sortOrder = sort.order.map(order => order.toLowerCase())
 
-  return _.orderBy(nodes, sortFields, sortOrder)
+  if (stats) {
+    sortFields.forEach(sortField => {
+      stats.uniqueSorts.add(sortField)
+    })
+  }
+
+  return _.orderBy(nodes, sortFns, sortOrder)
 }